refactor: Extract updateContractGroups to facilitate tests
