+++ src/helpers/groups.js
@@ -1,3 +1,5 @@
+import differenceBy from 'lodash/differenceBy'
+
 export const isExistingGroup = (groupsAlreadyCreated, groupToCreate) => {
   const isNameAlreadyUsed =
     groupsAlreadyCreated.find(
@@ -5,4 +7,15 @@ export const isExistingGroup = (groupsAlreadyCreated, groupToCreate) => {
     ) !== undefined
 
   return isNameAlreadyUsed
+}
+
+export const updateContactGroups = (contact, nextGroups) => {
+  const currentGroups = contact.groups.data
+  const toAdd = differenceBy(nextGroups, currentGroups, '_id')
+  const toRemove = differenceBy(currentGroups, nextGroups, '_id')
+
+  if (toAdd.length > 0) contact.groups.addById(toAdd.map(({ _id }) => _id))
+  else if (toRemove.length > 0)
+    contact.groups.removeById(toRemove.map(({ _id }) => _id))
+  // we can't do both at the same time right now, see https://github.com/cozy/cozy-client/issues/358
 }
